api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-docx
python-pptx
pymupdf
pytesseract
pillow
requests
openai
gTTS
